+++ src/Moon.AspNetCore/HttpErrorMessage.cs
@@ -3,10 +3,10 @@
     /// <summary>
     /// The HTTP error message.
     /// </summary>
-    public class ErrorMessage
+    public class HttpErrorMessage
     {
         /// <summary>
-        /// Gets or sets the error message that will be returned to user.
+        /// Gets or sets the message that will be returned to user.
         /// </summary>
         public string Message { get; set; }
     }